requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.7.2
astroid==2.15.6
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.3.2
cryptography==41.0.5
defusedxml==0.8.0rc2
dill==0.3.7
discord-webhook==1.3.0
Django==4.2.5
django-cors-headers==4.3.0
django-environ==0.11.2
elastic-transport==8.4.0
elasticsearch==8.9.0
elasticsearch-django==8.4
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
multidict==6.0.4
oauthlib==3.2.2
openai==0.28.1
openapi-codec==1.3.2
packaging==23.1
passlib==1.7.4
pip-autoremove==0.10.0
platformdirs==3.10.0
psycopg==3.1.12
psycopg2-binary==2.9.8
pycparser==2.21
PyJWT==2.8.0
pylint==2.17.5
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
python3-openid==3.2.0
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
simplejson==3.19.1
social-auth-app-django==5.4.0
social-auth-core==4.4.2
sqlparse==0.4.4
tiktoken==0.5.1
tomlkit==0.12.1
tqdm==4.66.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==1.26.18
whitenoise==6.5.0
wrapt==1.15.0
yarl==1.9.2
